+++ CHANGES.rst
@@ -1,22 +1,34 @@
 Changelog
 =========
 
-2.0.2 (unreleased)
+2.1.1 (unreleased)
 ------------------
-
-New Features:
 
 - Translate action name, workflow state and transition names in @history endpoint.
   [erral]
 
+- Translate FTI titles on `@types` endpoint. Fixes #337.
+  [erral]
+
+Bugfixes:
+
+- Make sure DX DefaultFieldDeserializer validates field values.
+  [lgraf]
+
+
+2.1.0 (2018-06-23)
+------------------
+
+New Features:
+
 - Include translated role title in `@roles` GET.
   [lgraf]
+
 
 2.0.1 (2018-06-22)
 ------------------
 
-- Translate FTI titles on `@types` endpoint. Fixes #337.
-  [erral]
+Bugfixes:
 
 - Hide upgrades from the add-ons control panel.
   Fixes `issue 532 <https://github.com/plone/plone.restapi/issues/532>`_.